refactor: decompose large components into hooks and sub-components
- Extract 7 custom hooks (useDigimonSearch, useTeamStorage,
  useEvolutionRequirements, useEvolutionTree, useNodePositioning,
  useEvolutionConnections, useVerticalEvolutionTree)
- Create shared DigimonSearchBar component (replaces 3x duplicate search UI)
- Split DigimonDetails into header, tab nav, path section, DNA partner
  display, and DNA requirements box sub-components
- Split MyTeam into TeamHeader, TeamSlot, and TeamGrid sub-components
- Move each component to a folder with index.tsx for cleaner sub-component
  co-location; import paths in App.tsx are unchanged
- Centralize ARMOR_DIGIMON_IDS (was duplicated 3x) and stage gradient
  colors into constants/
- EvolutionTreeGraph useEffect (356 lines) replaced by three focused
  useMemo-based hooks

Total: ~1,950 lines across 4 files -> ~800 lines + 14 hooks + 15 sub-components
