+++ src/constants/stageColors.ts
@@ -1,0 +1,9 @@
+import { DigimonStage } from '../types/digimon';
+
+export const STAGE_GRADIENT_COLORS: Record<DigimonStage, string> = {
+  'In-Training': 'from-green-400 to-green-600',
+  'Rookie': 'from-blue-400 to-blue-600',
+  'Champion': 'from-yellow-400 to-yellow-600',
+  'Ultimate': 'from-purple-400 to-purple-600',
+  'Mega': 'from-red-400 to-red-600',
+};
